+++ src/utils/sources-tree/createTree.js
@@ -2,6 +2,7 @@
 
 import { createNode, createParentMap } from "./utils";
 import { collapseTree } from "./collapseTree";
+import { sortEntireTree } from "./sortTree";
 import { addToTree } from "./addToTree";
 
 import type { SourcesMap } from "../../reducers/types";
@@ -12,7 +13,7 @@ export function createTree(sources: SourcesMap, debuggeeUrl: string) {
     addToTree(uncollapsedTree, source, debuggeeUrl);
   }
 
-  const sourceTree = collapseTree(uncollapsedTree);
+  const sourceTree = sortEntireTree(collapseTree(uncollapsedTree), debuggeeUrl);
 
   return {
     uncollapsedTree,
