+++ src/utils/sources-tree/index.js
@@ -5,6 +5,7 @@
 
 export { formatTree } from "./formatTree";
 export { addToTree } from "./addToTree";
+export { sortTree, sortEntireTree } from "./sortTree";
 export { collapseTree } from "./collapseTree";
 export { getDirectories } from "./getDirectories";
 export { createTree } from "./createTree";
